Fix for compare when a single use function is used (#4)
The macro INTERNAL_CUTI_ASSERT_COMPARE uses the macro input 'bound'
for creation of an information string and is passed to another macro
IMPL_CUTI_ASSERT for the actual assertion.

If the caller used a 'one-time' function, e.g. 'Initialize' directly in
a comparison macro then this would not execute correctly. The 'Initialize' function
would be called twice, which situation depending might produce the wrong
outcome or an unwanted exception.

Changed macro to execute the bound and store its result (or copy it for P.O.D).
This return value is then asserted for bounds and used to construct an information string.